+++ tests/testthat/test-sb_api_connect.R
@@ -1,6 +1,10 @@
 describe("sb_api_connect()", {
   it("should error when API credentials are not available", {
-    .sb_env$api_available <- FALSE
+    withr::local_envvar(
+      SUPABASE_URL = "",
+      SUPABASE_ANON_KEY = "",
+      SUPABASE_PUBLISHABLE_KEY = ""
+    )
     expect_snapshot(sb_api_connect(), error = TRUE)
   })
 })